requirements.txt
# VaultUSB Python Dependencies for DietPi Bookworm
# Core web framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# Template engine
jinja2>=3.1.2

# File upload support
python-multipart>=0.0.6

# Database
sqlmodel>=0.0.14
aiosqlite>=0.19.0

# Cryptography (optimized for Bookworm)
pynacl>=1.5.0
argon2-cffi>=23.1.0
itsdangerous>=2.1.2
cryptography>=41.0.0

# 2FA support
pyotp>=2.9.0

# Configuration
tomli>=2.0.1

# System monitoring
psutil>=5.9.6

# WebSocket support
websockets>=12.0

# DietPi Bookworm specific packages
setuptools>=68.0.0
wheel>=0.41.0
